+++ src/icemodel/IceModel.cc
@@ -57,6 +57,7 @@
 #include "pism/util/iceModelVec2T.hh"
 #include "pism/fracturedensity/FractureDensity.hh"
 #include "pism/coupler/util/options.hh" // ForcingOptions
+#include "pism/coupler/util/ScalarForcing.hh"
 
 namespace pism {